Move map controls to MapControls component
